server part structure changed
