--- node-server/api/v1/queries.js
@@ -1,22 +1,0 @@
-const Pool = require('pg').Pool;
-
-const pool = new Pool({
-  user: 'postgres',
-  host: 'localhost',
-  database: 'am',
-  password: '12345',
-  port: 5432
-});
-
-const getUsers = (request, response) => {
-  pool.query('SELECT * FROM users', (error, results) => {
-    if (error) {
-      throw error;
-    }
-    response.status(200).json(results.rows);
-  });
-};
-
-module.exports = {
-  getUsers
-};
